chore(inspector): refactor inspector2_findings_exist check into two
