+++ prowler/providers/aws/services/inspector2/inspector2_is_enabled/inspector2_is_enabled.metadata.json
@@ -1,14 +1,14 @@
 {
   "Provider": "aws",
-  "CheckID": "inspector2_findings_exist",
-  "CheckTitle": "Check if Inspector2 findings exist",
+  "CheckID": "inspector2_is_enabled",
+  "CheckTitle": "Check if Inspector2 is enabled",
   "CheckType": [],
   "ServiceName": "inspector2",
   "SubServiceName": "",
   "ResourceIdTemplate": "arn:aws:inspector2:region:account-id/detector-id",
   "Severity": "medium",
   "ResourceType": "Other",
-  "Description": "Check if Inspector2 findings exist",
+  "Description": "Check if Inspector2 is enabled",
   "Risk": "Without using AWS Inspector, you may not be aware of all the security vulnerabilities in your AWS resources, which could lead to unauthorized access, data breaches, or other security incidents.",
   "RelatedUrl": "https://docs.aws.amazon.com/inspector/latest/user/findings-understanding.html",
   "Remediation": {
